Final (?) fix for fix intended in 5b67f79daa363c52fd3b7c8dff9c2626001c80c4
It turns out that for value computation we only need to store the return sites to which we return in an unbalanced way. Then, at these sites, we *always* look up the edge function that goes to this site from the method's start point and the ZERO node! Previously, we did the look not using the ZERO node but using the original source of the path edge in the callee, which is wrong. Fortunately, the fix also makes for an easier implementation.