Adds method to debug IFDS edges
Extends the IDE solver to optionally record all intermediate edges created
by the user specified flow functions. These results can be used by
subclasses to debug or visualize the results of the algorithm.

Recording is opt-in, and is specified by the problem instance passed to the
solver. For backwards compatibility a new interface, DebugSolverConfiguration
has been added which currently contains the function: recordEdges which
indicates whether edges should be saved. For backwards compatibility, if the
problem does not implement the interface recordEdges is assumed to be
false.

It is unclear if the fieldsens solver could also be extended, but it has the
Debugger object which maybe (?) could be updated to do something similar.
It's possible that it is desirable to extend the current IDESolver to support
a similar mechanism, which would simply subsume this commit.