Adds dot export, expands solver debug configuration
In light of the discussion on the PR, I have decided to keep the
new interface in order to preserve backwards compatibility.
While reintroducing this change (which I had removed before
reversing my decision) I noticed that the IFDSSolver did not
pass along the information that the IFDS problem wanted to have
the edges recorded. Fixing this required pulling the anonymous inner
class created by the IFDSSolver into a static inner class (because
this object had to now implement two interfaces, which cannot be done
with anonymous classes).

I also have included the Dot export which led me to add this functionality
to begin with. I have genericized it: users now must provide a class
to handle creating a string representation of the methods, nodes, and facts
in the tabulation problem.