Better handling of interprocedural flow edges
Previously, we used the containing method of a unit to determine
whether or not a flow edge was interprocedural (roughly, if the
methods of the two endpoints were not equal the edge was judged
to be interprocedural).

This was fine, but missed recursive method calls. To get around this,
we simply save the interprocedural edges and intraprocedural edges
in different tables.

This required changing how we compute the dot file and so there is
more abstraction to reduce code duplication and cut down on complexity.

Finally: this commit reverts a mistaken change to the classpath
adds a default implementation of recordEdges in the default problem
instance, adds name to header comments, and removes the problematic
try-with-resource.