Allows the programmer to hook into value computation
Adds a new method, joinValueAt which allows the programmer to
hook into the value computation phase. The goal here is not to
change how values are computed (it is wholly expected that the
valueLattice will be used) but to allow the programmer to identify
unit/fact pairs of interest during the value computation phase.