Adds method filtering for the flow printer
Using the dot export on a large project is infeasible: although
the export function completes just fine the resulting dot file is
too large for dot to handle in any reasonable amount of time. This
commit allows the programmer to focus on a certain set of methods of
interest. For calls into/out-of these methods, the interprocedural
edges are still printed but the full bodies of the callee/called
methods are omitted and only the call/return nodes are included.